Implement Dashboard with Bottom Navigation and refactor feature structure
This commit introduces a main dashboard activity utilizing `BottomNavigationView` and Jetpack Navigation, while reorganizing meal-related logic into a more specific `home` feature module.

- **Dashboard & Navigation**:
    - Created `DashboardActivity` with `BottomNavigationView` and `NavHostFragment`.
    - Defined `dashboard_graph.xml` with five main destinations: Home, Search, Favourites, Plan, and Profile.
    - Implemented view fragments for each destination (`HomeFragment`, `SearchFragment`, etc.) with placeholder layouts.
    - Added custom navigation assets (icons for home, search, favorites, plan, and profile).
    - Configured `DashboardActivity` as the new launcher activity in `AndroidManifest.xml`.

- **Refactoring & Package Reorganization**:
    - Migrated meal-related DTOs, mappers, entities, and models from `features.meals` to `features.home`.
    - Updated `MealEntity` table name to `favMeals`.
    - Cleaned up empty placeholder files and removed unused `CategoryEntity` and its mapper.
    - Updated `LoginFragment` and `RegisterFragment` to navigate to `DashboardActivity` upon success.

- **UI Resources**:
    - Added `bottom_nav_menu.xml` and styled the navigation bar with rounded corners and primary color indicators.
    - Defined new color and string resources for the navigation items.
    - Created `styles.xml` to customize the Material 3 `BottomNavIndicator`.
