+++ app/src/main/java/com/example/recipe_android_project/features/dashboard/view/DashboardActivity.java
@@ -1,0 +1,42 @@
+package com.example.recipe_android_project.features.dashboard.view;
+
+import android.os.Bundle;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.navigation.NavController;
+import androidx.navigation.fragment.NavHostFragment;
+import androidx.navigation.ui.NavigationUI;
+
+import com.example.recipe_android_project.R;
+import com.google.android.material.bottomnavigation.BottomNavigationView;
+
+public class DashboardActivity extends AppCompatActivity {
+
+    private NavController navController;
+    private BottomNavigationView bottomNavigationView;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_dashboard);
+
+        setupNavigation();
+    }
+
+    private void setupNavigation() {
+        bottomNavigationView = findViewById(R.id.bottom_navigation);
+
+        NavHostFragment navHostFragment = (NavHostFragment) getSupportFragmentManager()
+                .findFragmentById(R.id.nav_host_fragment);
+
+        if (navHostFragment != null) {
+            navController = navHostFragment.getNavController();
+
+            NavigationUI.setupWithNavController(bottomNavigationView, navController);
+        }
+    }
+
+    @Override
+    public boolean onSupportNavigateUp() {
+        return navController.navigateUp() || super.onSupportNavigateUp();
+    }
+}
