--- app/src/main/java/com/example/recipe_android_project/features/meals/data/remote/api/api.java
@@ -1,4 +1,0 @@
-package com.example.recipe_android_project.features.meals.data.remote.api;
-
-public class api {
-}
