--- app/src/main/java/com/example/recipe_android_project/features/meals/domain/usecases/usecase.java
@@ -1,4 +1,0 @@
-package com.example.recipe_android_project.features.meals.domain.usecases;
-
-public class usecase {
-}
