--- app/src/main/java/com/example/recipe_android_project/features/meals/presentation/base/base.java
@@ -1,4 +1,0 @@
-package com.example.recipe_android_project.features.meals.presentation.base;
-
-public class base {
-}
