+++ app/src/main/res/drawable/home_icon.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="800dp"
+    android:height="800dp"
+    android:viewportWidth="32"
+    android:viewportHeight="32">
+  <path
+      android:pathData="M16,2.672l-5.331,5.331v-2.133h-4.265v6.398l-3.755,3.755 0.754,0.754 12.597,-12.597 12.597,12.597 0.754,-0.754 -13.351,-13.351zM7.47,6.937h2.132v2.132l-2.133,2.133v-4.265z"
+      android:fillColor="#000000"/>
+  <path
+      android:pathData="M6.404,16.533v12.795h7.464v-7.464h4.265v7.464h7.464v-12.795l-9.596,-9.596 -9.596,9.596zM24.53,28.262h-5.331v-7.464h-6.398v7.464h-5.331v-11.287l8.53,-8.53 8.53,8.53v11.287z"
+      android:fillColor="#000000"/>
+</vector>
