+++ app/src/main/res/drawable/plan_icon.xml
@@ -1,0 +1,11 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="800dp"
+    android:height="800dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M18,4L18,0L18,4ZM7,18L5,18L7,18ZM19,18L9,18L19,18ZM7,14L5,14L7,14ZM19,14L9,14L19,14ZM6,4L6,0L6,4ZM1,9L23,9L1,9ZM1,23L23,23L23,4L1,4L1,23Z"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"/>
+</vector>
