+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <!-- IMPORTANT: Menu item IDs MUST match fragment IDs in nav_graph.xml -->
+
+    <item
+        android:id="@+id/homeFragment"
+        android:icon="@drawable/home_icon"
+        android:title="@string/home" />
+
+    <item
+        android:id="@+id/searchFragment"
+        android:icon="@drawable/search_icon"
+        android:title="@string/search" />
+
+    <item
+        android:id="@+id/favouriteFragment"
+        android:icon="@drawable/favorite_icon"
+        android:title="@string/favourites" />
+
+    <item
+        android:id="@+id/planFragment"
+        android:icon="@drawable/plan_icon"
+        android:title="@string/plan" />
+
+    <item
+        android:id="@+id/profileFragment"
+        android:icon="@drawable/profile_icon"
+        android:title="@string/profile" />
+
+</menu>
