+++ app/src/main/res/values/colors.xml
@@ -16,6 +16,7 @@
     <color name="text_secondary">#8A8A8E</color>
     <color name="text_hint">#9AA0A6</color>
     <color name="colorGray">#BDBDBD</color>
+    <color name="gray">#9E9E9E</color>
 
     <!-- Status Colors -->
     <color name="error">#E53935</color>
